Moving under each js page file page/ folder for easier minify
