+++ GapsWeb/src/main/resources/static/js/page/configuration.js
@@ -1,27 +1,28 @@
 /*
- * Copyright 2019 Jason H House
+ * Copyright 2020 Jason H House
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ *  The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
  *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
  */
 
 /* global Handlebars, SockJS, Stomp */
 /* eslint no-undef: "error" */
 
-import hideAllAlertsAndSpinners from './modules/alerts-manager.min.js';
-import saveSchedule from './modules/schedule.min.js';
-import Payload from './modules/payload.min.js';
-import { saveTelegramNotifications, testTelegramNotifications } from './modules/telegram-notifications.min.js';
-import { saveSlackNotifications, testSlackNotifications } from './modules/slack-notifications.min.js';
-import { savePushBulletNotifications, testPushBulletNotifications } from './modules/push-bullet-notifications.min.js';
-import { saveGotifyNotifications, testGotifyNotifications } from './modules/gotify-notifications.min.js';
-import { saveEmailNotifications, testEmailNotifications } from './modules/email-notifications.min.js';
-import { saveDiscordNotifications, testDiscordNotifications } from './modules/discord-notifications.min.js';
-import { getSoundOptions, savePushOverNotifications, testPushOverNotifications } from './modules/push-over-notifications.min.js';
-import { openPlexLibraryConfigurationModel, savePlexLibraryConfiguration } from './modules/plex-configuration.min.js';
+import hideAllAlertsAndSpinners from '../modules/alerts-manager.min.js';
+import saveSchedule from '../modules/schedule.min.js';
+import Payload from '../modules/payload.min.js';
+import { saveTelegramNotifications, testTelegramNotifications } from '../modules/telegram-notifications.min.js';
+import { saveSlackNotifications, testSlackNotifications } from '../modules/slack-notifications.min.js';
+import { savePushBulletNotifications, testPushBulletNotifications } from '../modules/push-bullet-notifications.min.js';
+import { saveGotifyNotifications, testGotifyNotifications } from '../modules/gotify-notifications.min.js';
+import { saveEmailNotifications, testEmailNotifications } from '../modules/email-notifications.min.js';
+import { saveDiscordNotifications, testDiscordNotifications } from '../modules/discord-notifications.min.js';
+import { getSoundOptions, savePushOverNotifications, testPushOverNotifications } from '../modules/push-over-notifications.min.js';
+import { openPlexLibraryConfigurationModel, savePlexLibraryConfiguration } from '../modules/plex-configuration.min.js';
 
 function testTmdbKey() {
   hideAllAlertsAndSpinners();
