+++ GapsWeb/src/main/resources/static/js/page/index.js
