+++ GapsWeb/src/main/resources/templates/configuration.html
@@ -443,6 +443,6 @@
 <script src="/js/libraries/sockjs-1.4.0.min.js" type="text/javascript"></script>
 <script src="/js/libraries/stomp-2.3.3.min.js" type="text/javascript"></script>
 <script src="/js/libraries/handlebars-v4.7.6.min.js" type="text/javascript"></script>
-<script src="/js/configuration.min.js" type="module"></script>
+<script src="/js/page/configuration.min.js" type="module"></script>
 </body>
 </html>
