+++ GapsWeb/src/main/resources/templates/libraries.html
@@ -184,6 +184,6 @@
 <script src="/js/libraries/bootstrap.bundle.min.js" type="text/javascript"></script>
 <script src="/js/libraries/datatables.min.js" type="text/javascript"></script>
 <script src="/js/libraries/handlebars-v4.7.6.min.js" type="text/javascript"></script>
-<script type="module" src="/js/libraries.min.js"></script>
+<script type="module" src="/js/page/libraries.min.js"></script>
 </body>
 </html>
