+++ GapsWeb/src/main/resources/templates/login.html
@@ -51,6 +51,6 @@
 
 <script src="/js/libraries/jquery-3.4.1.min.js" type="text/javascript"></script>
 <script src="/js/libraries/bootstrap.bundle.min.js" type="text/javascript"></script>
-<script src="/js/index.min.js" type="text/javascript"></script>
+<script src="/js/page/index.min.js" type="text/javascript"></script>
 </body>
 </html>
