+++ GapsWeb/src/main/resources/templates/mislabeled.html
@@ -94,6 +94,6 @@
 <script src="/js/libraries/jquery-3.4.1.min.js" type="text/javascript"></script>
 <script src="/js/libraries/bootstrap.bundle.min.js" type="text/javascript"></script>
 <script src="/js/libraries/datatables.min.js" type="text/javascript"></script>
-<script src="/js/mislabeled.min.js" type="text/javascript"></script>
+<script src="/js/page/mislabeled.min.js" type="text/javascript"></script>
 </body>
 </html>
