+++ GapsWeb/src/main/resources/templates/recommended.html
@@ -218,6 +218,6 @@
 <script src="/js/libraries/sockjs-1.4.0.min.js" type="text/javascript"></script>
 <script src="/js/libraries/stomp-2.3.3.min.js" type="text/javascript"></script>
 <script src="/js/libraries/handlebars-v4.7.6.min.js" type="text/javascript"></script>
-<script type="module" src="/js/recommended.min.js"></script>
+<script type="module" src="/js/page/recommended.min.js"></script>
 </body>
 </html>
